+++ CHANGELOG.md
@@ -79,6 +79,8 @@ for specific instructions.
   deprecated in favor of `metrics`. Flag names starting with `prometheus.` have
   also been deprecated in favor of the same flags with the `metrics.` prefix.
   (@rfratto)
+  
+- [DEPRECATION] Rename Tempo to Traces (@mattdurham)
 
 # v0.18.4 (2021-09-14)